Add a README
Why - because GitHub Actions is being weird about re-running its
workflow when it ran the latest build with bizarrely out of date code.

I suspect that when I pushed multiple commits, that it built off the
first of the commits to be pushed to GitHub, rather than the latest
commit out of the batch.